Fixing version number
At some point the version number was switched from 2.7_mb2.1 to just mb2.1,
I think it had something to do with how we were building under CMake and
setuptools not distinguishing between _ and -, but I'm not sure.

Unfortunately, every ----ing repo hardcodes the name of the pyserial egg,
and it seems like we can build fine now with the old version number, so rather
than switching all the other repos, I'm switching this back.